Buffer data until len(data) is a multiple of 8.
Windows (and probably other OS's too) will send partial chunks of data
over bluetooth to avoid excess CPU load. But to convert to an array of
doubles the data chunk needs to have a length that is a multiple of
eight. So, we can just keep on appending data until that limit is
reached, and the DSP math works out just the same.